+++ showcase/modules/documentation/elements/icon/code/1-icon.tsx
@@ -1,10 +1,13 @@
 import * as React from "react";
 import { Icon } from "bambu/lib/elements/icon";
+import FaIcon from "bambu/lib/elements/icon/FaIcon";
 
 export default class Index extends React.Component<{}, {}> {
     public render() {
         return (
-            <Icon name="home" />
+            <Icon>
+                <FaIcon name="home" />
+            </Icon>
         );
     }
 }