+++ showcase/modules/documentation/elements/icon/code/2-colors.tsx
@@ -1,15 +1,15 @@
 import * as React from "react";
-import { Icon } from "bambu/lib/elements/icon";
+import {FaIcon, Icon} from "bambu/lib/elements/icon";
 import bulma from "bambu/lib/base/css/bulma";
 
 export default class Index extends React.Component<{}, {}> {
     public render() {
         return (
            <div>
-               <Icon name="info-circle" className={bulma.hasTextInfo} />
-               <Icon name="check-square" className={bulma.hasTextSuccess} />
-               <Icon name="exclamation-triangle" className={bulma.hasTextWarning} />
-               <Icon name="ban" className={bulma.hasTextDanger} />
+               <Icon className={bulma.hasTextInfo}><FaIcon name="info-circle" /></Icon>
+               <Icon className={bulma.hasTextSuccess}><FaIcon name="check-square"  /></Icon>
+               <Icon className={bulma.hasTextWarning}><FaIcon name="exclamation-triangle" /></Icon>
+               <Icon className={bulma.hasTextDanger}><FaIcon name="ban" /></Icon>
            </div>
         );
     }