+++ showcase/modules/documentation/elements/icon/code/3-stack.tsx
@@ -1,16 +1,15 @@
 import * as React from "react";
-import {Icon} from "bambu/lib/elements/icon";
+import {FaIcon, Icon} from "bambu/lib/elements/icon";
 import {Size} from "bambu";
-import {IconStack} from "bambu/lib/elements/icon/FaIcon";
-import IconStacker from "bambu/lib/elements/icon/IconStacker";
+import {FaIconStack} from "bambu/lib/elements/icon/FaIcon";
 
 export default class Index extends React.Component<{}, {}> {
     public render() {
         return (
-           <IconStacker>
-               <Icon name="circle" bSize={Size.medium} stack={IconStack.x2} />
-               <Icon name="flag" bSize={Size.medium} stack={IconStack.x1} inverse />
-           </IconStacker>
+            <Icon stack bSize={Size.medium}>
+                <FaIcon name="circle" stack={FaIconStack.x2} />
+                <FaIcon name="flag" stack={FaIconStack.x1} inverse />
+            </Icon>
         );
     }
 }