+++ src/elements/icon/index.ts
@@ -1,4 +1,17 @@
 export {default as Icon, IconProps, IconOptions} from './Icon';
-export {default as FaIcon, FaIconProps, IconSize, iconSizeValues, IconStack} from './FaIcon';
+export {
+    default as FaIcon,
+    FaIconProps,
+    FaIconSize,
+    faIconSizeValues,
+    FaIconStack,
+    FaIconEffect,
+    faIconEffectValues,
+    FaIconFlip,
+    faIconFlipValues,
+    faIconStackValues,
+    FaIconStyle,
+    faIconStyleValues
+} from './FaIcon';
+export {default as MaIcon, MaIconFlip, MaIconColor, maIconFlipValues, MaIconProps, MaIconSize} from "./MaIcon";
 export {default as FaIconList } from "./FaIconList";
-export {default as IconStacker } from "./IconStacker";