+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 * added *Layer* component to show children as vertical parts which are  wrapped with *Content* in a *Box* component.
 * added finding modules in bambu and generating modules json object.
 * changed `elementRef` property type as function.
+* removed *FaIcon* from *Icon* component. You should give it as a children.
+* removed *MaIcon* component. https://materialdesignicons.com/getting-started
+
 
 #### v0.6.201-beta.2
 * solved https://github.com/wasabi-io/bambu/issues/6